+++ app/views/humas/bkk/mou/index.php
@@ -10,12 +10,10 @@
 
     <div class="row mb-4 ">
         <div class="col-lg-6">
-            <button type="button" class="btn btn-primary tombolTambahDataMou" data-bs-toggle="modal"
-                data-bs-target="#formModal">
+            <button type="button" class="btn btn-primary tombolTambahDataMou" data-bs-toggle="modal" data-bs-target="#formModal">
                 Tambah Data MoU
             </button>
-            <button type="button" class="btn btn-primary my-3 mx-3 tampilModalImport" data-toggle="modal"
-                data-target="#modalImport">
+            <button type="button" class="btn btn-primary my-3 mx-3 tampilModalImport" data-toggle="modal" data-target="#modalImport">
                 Import Data Dari Excel
             </button>
         </div>
@@ -33,7 +31,7 @@
     </div> -->
 
 
-    <table id="myTable" class="table table-striped" style="width:100%">
+    <table id="table" class="table table-striped" style="width:100%">
         <thead>
             <tr>
                 <th>Aksi</th>
@@ -45,21 +43,18 @@
         </thead>
         <tbody>
             <?php $i = 1;
-      foreach ($data['siswa'] as $siswa) : ?>
-            <tr>
-                <td>
-                    <a href="<?= BASEURL; ?>/bkk/ubahmou/<?= $siswa['id']; ?>"
-                        class="badge text-bg-warning tampilModalEditdas" style="text-decoration: none;"
-                        data-bs-toggle="modal" data-bs-target="#formModal" data-id="<?= $siswa['id']; ?>"> Edit</a>
-                    <a href="<?= BASEURL; ?>/bkk/hapusmou/<?= $siswa['id']; ?>" class="badge text-bg-danger"
-                        style="text-decoration: none;" onclick="return confirm('Yakin ingin menghapus?');"> Hapus</a>
-                </td>
-                <td><?= $i++ ?></td>
-                <td><?= $siswa['namaperusahaan']; ?></td>
-                <td><?= $siswa['tanggalmou']; ?></td>
-                <td><?= $siswa['updraftmou']; ?></td>
+            foreach ($data['siswa'] as $siswa) : ?>
+                <tr>
+                    <td>
+                        <a href="<?= BASEURL; ?>/bkk/ubahmou/<?= $siswa['id']; ?>" class="badge text-bg-warning tampilModalEditdas" style="text-decoration: none;" data-bs-toggle="modal" data-bs-target="#formModal" data-id="<?= $siswa['id']; ?>"> Edit</a>
+                        <a href="<?= BASEURL; ?>/bkk/hapusmou/<?= $siswa['id']; ?>" class="badge text-bg-danger" style="text-decoration: none;" onclick="return confirm('Yakin ingin menghapus?');"> Hapus</a>
+                    </td>
+                    <td><?= $i++ ?></td>
+                    <td><?= $siswa['namaperusahaan']; ?></td>
+                    <td><?= $siswa['tanggalmou']; ?></td>
+                    <td><?= $siswa['updraftmou']; ?></td>
 
-            </tr>
+                </tr>
             <?php endforeach; ?>
         </tbody>
     </table>
@@ -103,20 +98,17 @@
                     <input type="hidden" name="id" id="id">
                     <div class="mb-3">
                         <label for="namaperusahaan" class="form-label">Nama Perusahaan</label>
-                        <input type="text" class="form-control" id="namaperusahaan" name="namaperusahaan"
-                            placeholder="Masukkan nama perusahaan" autocomplete="off">
+                        <input type="text" class="form-control" id="namaperusahaan" name="namaperusahaan" placeholder="Masukkan nama perusahaan" autocomplete="off">
                     </div>
 
                     <div class="mb-3">
                         <label for="tanggalmou" class="form-label">Tanggal MoU</label>
-                        <input type="text" class="form-control" id="tanggalmou" name="tanggalmou"
-                            placeholder="Masukkan tanggal" autocomplete="off">
+                        <input type="text" class="form-control" id="tanggalmou" name="tanggalmou" placeholder="Masukkan tanggal" autocomplete="off">
                     </div>
 
                     <div class="mb-3">
                         <label for="updraftmou" class="form-label">Upload Draft MoU</label>
-                        <input type="text" class="form-control" id="updraftmou" name="updraftmou"
-                            placeholder="Masukkan updraftmou" autocomplete="off">
+                        <input type="text" class="form-control" id="updraftmou" name="updraftmou" placeholder="Masukkan updraftmou" autocomplete="off">
                     </div>
